+++ CHANGELOG.md
@@ -9,6 +9,9 @@ For experimental package changes, see the [experimental CHANGELOG](experimental/
 
 ### :boom: Breaking Change
 
+* feat(otlp-exporter-base)!: added support for sending gzip compressed content in the OTLP exporter for browser [#4493](https://github.com/open-telemetry/opentelemetry-js/pull/4493)
+  * The function signature for `sendWithXhr` has changed.
+
 ### :rocket: (Enhancement)
 
 * perf(sdk-trace-base): do not allocate arrays if resource has no pending async attributes